requirements.txt
flask
pypdf
python-pptx
python-docx
openpyxl
requests
youtube-transcript-api